Fix compatibility tests>simple debuggable examples
