+++ compatibility-tests/test-parcel1-react17/package.json
@@ -4,12 +4,12 @@
     "dev": "parcel serve ./index.html"
   },
   "dependencies": {
-    "parcel-bundler": "^1.12.5",
     "@react-three/drei": "^7.3.1",
     "@react-three/fiber": "^7.0.6",
     "@theatre/core": "^0.0.1-COMPAT.1",
     "@theatre/r3f": "^0.0.1-COMPAT.1",
     "@theatre/studio": "^0.0.1-COMPAT.1",
+    "parcel-bundler": "^1.12.5",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "three": "^0.137.0"
