+++ compatibility-tests/test-parcel1-react18/package.json
@@ -4,8 +4,9 @@
     "dev": "parcel serve ./index.html"
   },
   "dependencies": {
-    "@theatre/core": "0.0.1-COMPAT.1",
-    "@theatre/studio": "0.0.1-COMPAT.1",
+    "@theatre/core": "^0.0.1-COMPAT.1",
+    "@theatre/r3f": "^0.0.1-COMPAT.1",
+    "@theatre/studio": "^0.0.1-COMPAT.1",
     "parcel-bundler": "^1.12.5",
     "react": "^18.1.0",
     "react-dom": "^18.1.0"
