--- compatibility-tests/test-parcel1-react18/src/App.js
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export default function App({obj}) {
-  console.log(obj)
-  return <div>hi</div>
-}
