+++ packages/r3f/package.json
@@ -62,7 +62,7 @@
     "react-shadow": "^19.0.3",
     "react-use-measure": "^2.0.4",
     "reakit": "^1.3.8",
-    "styled-components": "^5.3.0",
+    "styled-components": "^5.3.5",
     "typescript": "^4.4.2",
     "zustand": "^3.5.1"
   },
